[WT-1239] refactor wallet-allowed list to wallet  folder
