+++ packages/checkout/sdk/src/connect/connect.ts
@@ -5,14 +5,9 @@ import {
   ConnectParams,
   WALLET_ACTION,
   CheckConnectionResult,
-  GetWalletAllowListParams,
-  WalletFilterTypes,
-  GetWalletAllowListResult,
-  WalletInfo,
 } from '../types';
 import { Web3Provider, ExternalProvider } from '@ethersproject/providers';
 import { CheckoutError, CheckoutErrorType, withCheckoutError } from '../errors';
-import masterWalletList from './wallet_master_list.json';
 
 export async function checkIsWalletConnected(
   providerPreference: ConnectionProviders
@@ -97,31 +92,4 @@ async function getMetaMaskProvider(): Promise<Web3Provider> {
     );
 
   return new Web3Provider(provider);
-}
-
-export async function getWalletAllowList({
-  type = WalletFilterTypes.ALL,
-  exclude,
-}: GetWalletAllowListParams): Promise<GetWalletAllowListResult> {
-  // todo:For async API call, use the CheckoutError with errorType:API_CALL_ERROR?? or any other
-
-  const filteredWalletsList = masterWalletList
-    .filter((wallet) => {
-      const walletNotExcluded = !exclude
-        ?.map((excludeWallet) => excludeWallet.connectionProvider)
-        .includes(wallet.connectionProvider as ConnectionProviders);
-
-      const allowAllWallets = type === WalletFilterTypes.ALL;
-      const walletsAllowedForType = wallet.platform.includes(type);
-
-      return walletNotExcluded && (allowAllWallets || walletsAllowedForType);
-    })
-    .map((wallet) => {
-      const { platform, ...walletInfo } = wallet;
-      return walletInfo as WalletInfo;
-    });
-
-  return {
-    allowedWallets: filteredWalletsList,
-  };
 }
