+++ packages/checkout/sdk/src/wallet/index.ts
@@ -1,0 +1,1 @@
+export * from './wallet';
